requirements.txt
numpy==1.26.4
pandas==2.2.0
torch
tensorflow
scikit-learn==1.6.1
langchain-core
langchain-groq
python-dotenv
